# lib/api-zod/src/generated/types/sendOpenaiMessageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SendOpenaiMessageBody {
  content: string;
  userName?: string;
  userAge?: number;
  userGender?: string;
}