# lib/api-zod/src/generated/types/textToSpeechBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TextToSpeechBodyVoice } from "./textToSpeechBodyVoice";

export interface TextToSpeechBody {
  text: string;
  voice: TextToSpeechBodyVoice;
}