# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: number;
  name: string;
  age: number;
  gender: string;
  conversationCount: number;
  rank: string;
  preferredVoice: string;
  createdAt: Date;
}